+++ src/services/openWeatherOneCall/axios.ts
@@ -3,8 +3,8 @@ import Config from 'react-native-config';
 import { getLocalizedTextFromMap } from '@libs/utils/localize/localize';
 import HttpClient from '@services/httpClient/httpClient';
 import { HTTP_CLIENT_STATUS_LIST } from '@services/httpClient/httpClient.const';
-import { getHttpClientStatusMessage, httpClientError } from '@services/httpClient/utils';
-import { OPEN_WEATHER_ONE_CALL_SERVICE_ERROR_STATUS } from '@services/openWeatherOneCall/status';
+import { getHttpClientStatusMessage, httpClientError } from '@services/httpClient/httpClient.util';
+import { OPEN_WEATHER_ONE_CALL_SERVICE_ERROR_STATUS } from '@services/openWeatherOneCall/axios.const';
 
 import type { HttpClientStatusList } from '@services/httpClient/httpClient.type';
 import type {
