+++ src/services/openWeatherOneCall/mock/test.mock.ts
@@ -1,5 +1,5 @@
 import { DEFAULT_LANGUAGE } from '@libs/utils/localize/localize.const';
-import { OPEN_WEATHER_ONE_CALL_SERVICE_ERROR_STATUS } from '@services/openWeatherOneCall/status';
+import { OPEN_WEATHER_ONE_CALL_SERVICE_ERROR_STATUS } from '@services/openWeatherOneCall/axios.const';
 
 import type { PickedAxiosResponse } from '@services/httpClient/httpClient.type';
 import type {
