+++ src/services/template/__test__/query.test.tsx
@@ -9,7 +9,7 @@ import { act, render, renderHook, screen, waitFor } from '@testing-library/react
 import ErrorBoundary from 'react-native-error-boundary';
 
 import { templateService } from '@services/template/axios';
-import { TEMPLATE_SERVICE_MOCK } from '@services/template/mock/mock';
+import { TEMPLATE_SERVICE_MOCK } from '@services/template/mock/test.mock';
 import { useGetTemplate } from '@services/template/query';
 
 // 서비스 모듈 모킹
